mm/compaction: use proper zoneid for compaction_suitable()
There are presently three compaction paths:

- direct compaction
- kcompactd compcation
- proc triggered compaction

When we do compaction in all these paths, we use compaction_suitable() to
check whether a zone is suitable to do compaction.

There're some issues around the usage of compaction_suitable().  We don't
use the proper zoneid in kcompactd_node_suitable() when try to wakeup
kcompactd.  In the kcompactd compaction paths, we call
compaction_suitable() twice and the zoneid isn't proper in the second
call.  For proc triggered compaction, the classzone_idx is always zero.

In order to fix these issues, I change the type of classzone_idx in the
struct compact_control from const int to int and assign the proper zoneid
before calling compact_zone().

This patch also fixes some comments in struct compact_control, as these
fields are not only for direct compactor but also for all other
compactors.

Link: http://lkml.kernel.org/r/1564062621-8105-1-git-send-email-laoar.shao@gmail.com
Fixes: ebff398017c6("mm, compaction: pass classzone_idx and alloc_flags to watermark checking")
Fixes: 698b1b30642f("mm, compaction: introduce kcompactd")
Signed-off-by: Yafang Shao <laoar.shao@gmail.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Paul Gortmaker <paul.gortmaker@windriver.com>
Cc: Rik van Riel <riel@redhat.com>
Cc: Yafang Shao <shaoyafang@didiglobal.com>
Cc: Mel Gorman <mgorman@techsingularity.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>