block: annotate refault stalls from IO submission
psi tracks the time tasks wait for refaulting pages to become uptodate,
but it does not track the time spent submitting the IO.  The submission
part can be significant if backing storage is contended or when cgroup
throttling (io.latency) is in effect - a lot of time is spent in
submit_bio().  In that case, we underreport memory pressure.

Annotate submit_bio() to account submission time as memory stall when the
bio is reading userspace workingset pages.

Link: http://lkml.kernel.org/r/20190808190300.GA9067@cmpxchg.org
Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Tested-by: Suren Baghdasaryan <surenb@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>