powperc/mm: read TLB Block Invalidate Characteristics
The PAPR document specifies the TLB Block Invalidate Characteristics which
tells for each pair of segment base page size, actual page size, the size
of the block the hcall H_BLOCK_REMOVE supports.

These characteristics are loaded at boot time in a new table hblkr_size.
The table is separate from the mmu_psize_def because this is specific to
the pseries platform.

A new init function, pseries_lpar_read_hblkrm_characteristics() is added to
read the characteristics. It is called from pSeries_setup_arch().

Fixes: ba2dd8a26baa ("powerpc/pseries/mm: call H_BLOCK_REMOVE")
Signed-off-by: Laurent Dufour <ldufour@linux.ibm.com>